can pass function to tag button
