+++ project/frontend/src/components/CategoryTagButtons.tsx
@@ -1,4 +1,3 @@
-import { Button } from 'react-bootstrap';
 import { CategoryTagButton } from './CategoryTagButton';
 
 /**
@@ -6,9 +5,9 @@ import { CategoryTagButton } from './CategoryTagButton';
  * @param tags is a list of strings that represents the tags for a category
  * @returns a collection of buttons that each represents one of theese tags
  */
-export function CategoryTagButtons(tags: string[]) : JSX.Element[] {
+export function CategoryTagButtons(tags: string[], filterCategory: ((arg0: string) => void)) : JSX.Element[] {
     const res = tags.map((t : string) => {
-        return CategoryTagButton(t)
+        return CategoryTagButton(t, filterCategory)
       })
     return res
 }
