+++ project/frontend/src/components/ShowEventCategories.tsx
@@ -7,33 +7,7 @@ import '../styles/showEventCat.css';
 import { EventCards} from './EventCards';
 import { category } from './Models';
 import { CategoryTagButtons } from './CategoryTagButtons';
-
-/**
- * Responsive item settings for different platform sizes.
- * Used in Carousel.
- */
-const responsive = {
-  desktopBig: {
-    breakpoint: { max: 3000, min: 1550 },
-    items: 4,
-    slidesToSlide: 4,
-  },
-  desktopSmall: {
-    breakpoint: { max: 1550, min: 1048 },
-    items: 3,
-    slidesToSlide: 3,
-  },
-  tablet: {
-    breakpoint: { max: 1048, min: 576 },
-    items: 2,
-    slidesToSlide: 2,
-  },
-  mobile: {
-    breakpoint: { max: 576, min: 0 },
-    items: 1,
-    slidesToSlide: 1,
-  }
-};
+import { CategoryItem } from './CategoryItem';
 
 /**
  * Retrieves data from backend using axios.
@@ -54,27 +28,14 @@ export default function DataCat() {
       const allCategories = res.data
       setAppState({ loading: false, cat: allCategories });
     })
-  }, [setAppState])
+  }, [])
+
+
 
   const categoryList = 
       <ListGroup>
       {appState.cat.map(c => (
-        <ListGroupItem>
-          <div className="category-title">
-            <h1 className="cat-title">{c.category}</h1>
-            <div className='cat-tags'>{CategoryTagButtons(c.tags) }</div>
-          </div>
-          
-          <Carousel 
-          responsive={responsive}
-          showDots={true}
-          /*infinite={true}*/
-          shouldResetAutoplay={false}
-          customTransition="1000ms cubic-bezier(0.645, 0.045, 0.355, 1) 0s"
-          >
-            {EventCards(c.events)}
-          </Carousel>
-        </ListGroupItem>
+        CategoryItem(c)
       ))}
     </ListGroup>
